refactor: split resolve-plan-run helpers under the CRAP ceiling
Move pure plan-run resolution into lib/orchestration so the new CLI stays
thin enough for the pre-push quality preview gate.
